pyproject.toml
[project]
name = "olapdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "tarcisio", email = "20189+tarcisio@users.noreply.github.com" }
]
requires-python = ">=3.14"
dependencies = [
    "pep249abc",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
]
